feat: refactor the code to diff file as helper method
